Move totals from common/components into dashboard/components
