+++ src/pages/dashboard/components/Totals.tsx
@@ -11,10 +11,10 @@
 import { Button } from '@invoiceninja/forms';
 import { AxiosResponse } from 'axios';
 import { endpoint } from 'common/helpers';
-import { Chart } from 'components/totals/Chart';
+import { Chart } from 'pages/dashboard/components/Chart';
 import { useEffect, useState } from 'react';
 import { Spinner } from 'components/Spinner';
-import { DropdownDateRangePicker } from '../DropdownDateRangePicker';
+import { DropdownDateRangePicker } from '../../../components/DropdownDateRangePicker';
 import { Card } from '@invoiceninja/cards';
 import { useTranslation } from 'react-i18next';
 import { InfoCard } from 'components/InfoCard';
